Split HomePage.cs
